requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
python-dotenv
faiss-cpu  
google-generativeai
numpy 
pandas